remove floats from text render target coords in `CTK_RenderText`
This caused fontsizes other than 11 to cause blurry awfulness...
I HATE FLOATS. WHY IS EVERY PIECE OF SDL RENDERING ASKING FOR FLOATS?

+++ CTK.h
@@ -3105,7 +3105,7 @@ CTK_RenderText(SDL_Renderer            *r,
 	int      text_w;
 	int      text_h;
 	SDL_Rect underline_r;
-	float    x, y;
+	int      x, y;
 
 	if (NULL == text ||
 	    NULL == text->text) {
